requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.12
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
pkg-resources==0.0.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
whitenoise==5.3.0
